requirements.txt
paho-mqtt
bson
pyaml